+++ packages/core/src/types.ts
@@ -82,14 +82,14 @@ export type FormDataKeys<T> = T extends Function | FormDataConvertibleValue
  */
 type ArrayFormDataKeys<T extends unknown[]> = number extends T['length']
   ? // Dynamic array
-    | `${number}`
+      | `${number}`
       | (0 extends 1 & T[number]
           ? never
           : T[number] extends FormDataConvertibleValue
             ? never
             : `${number}.${FormDataKeys<T[number]>}`)
   : // Tuple with known length
-    | Extract<keyof T, `${number}`>
+      | Extract<keyof T, `${number}`>
       | {
           [Key in Extract<keyof T, `${number}`>]: 0 extends 1 & T[Key]
             ? never
@@ -204,7 +204,7 @@ export type ScrollRegion = {
 export interface ClientSideVisitOptions<TProps = Page['props']> {
   component?: Page['component']
   url?: Page['url']
-  props?: ((props: TProps) => PageProps) | PageProps
+  props?: ((props: TProps, onceProps: Partial<TProps>) => PageProps) | PageProps
   flash?: ((flash: FlashData) => PageFlashData) | PageFlashData
   clearHistory?: Page['clearHistory']
   encryptHistory?: Page['encryptHistory']
@@ -550,30 +550,29 @@ export type InertiaAppConfig = {
   visitOptions?: (href: string, options: VisitOptions) => VisitOptions
 }
 
-export interface LinkComponentBaseProps
-  extends Partial<
-    Pick<
-      Visit<RequestPayload>,
-      | 'data'
-      | 'method'
-      | 'replace'
-      | 'preserveScroll'
-      | 'preserveState'
-      | 'preserveUrl'
-      | 'only'
-      | 'except'
-      | 'headers'
-      | 'queryStringArrayFormat'
-      | 'async'
-      | 'viewTransition'
-    > &
-      VisitCallbacks & {
-        href: string | UrlMethodPair
-        prefetch: boolean | LinkPrefetchOption | LinkPrefetchOption[]
-        cacheFor: CacheForOption | CacheForOption[]
-        cacheTags: string | string[]
-      }
-  > {}
+export interface LinkComponentBaseProps extends Partial<
+  Pick<
+    Visit<RequestPayload>,
+    | 'data'
+    | 'method'
+    | 'replace'
+    | 'preserveScroll'
+    | 'preserveState'
+    | 'preserveUrl'
+    | 'only'
+    | 'except'
+    | 'headers'
+    | 'queryStringArrayFormat'
+    | 'async'
+    | 'viewTransition'
+  > &
+    VisitCallbacks & {
+      href: string | UrlMethodPair
+      prefetch: boolean | LinkPrefetchOption | LinkPrefetchOption[]
+      cacheFor: CacheForOption | CacheForOption[]
+      cacheTags: string | string[]
+    }
+> {}
 
 type PrefetchObject = {
   params: ActiveVisit